fix: mark closets as notele in ossuary_zaba_flooded
This is a bit of an overkill / quick fix, in that (as far as I can tell)
closets generated by the flooding algorithm here are guaranteed to be
deep water, and so marking all deep water as notele seems to do the
trick. Probably it would be better if this algorithm didn't generate
closets at all, but someone would have to understand it to fix that.
Relevant mainly to species that can inhabit deep water of course.

Recently trapped a player in:
https://cbro.berotato.org/morgue/pigah/morgue-pigah-20220513-145335.txt