Despoiler the Abyss rune (mostly)
Replace external uses of the secret 'abyss rune tile' with a new
'runelight' feature, which stops autoexplore and is explicitly described
as indicating the presence of a nearby rune. Possible that some or all
of the internal tiles in these vaults should also be replaced with
runelights, for those with dungeon mapping mutations or similar.

Art is placeholder.

# Conflicts:
#	crawl-ref/source/dungeon-feature-type.h
#	crawl-ref/source/terrain.cc