feat: Remove an autoexplore trap (Lici, ebonnov)
This vault places two `8`s and two `9`s, and it does so in a way that's
likely to lead to autoexploring somewhat close to these monsters, due to
the corner at the entrance. Add a couple glass tiles to reveal the
innermost monsters (and loot) further before said corners are taken.

I considered other adjustments, but went with ebonnov's suggestion,
since it's as minimal as we're likely to get, and it's good to be very
minimal when adjusting for this sort of thing.