monster melee fixes
Slaying bonuses now apply to monsters without a weapon.

For monsters, +1 and +0 weapons did the same damage because random2(0) and random2(1) are both 0.

Monster max damage with a weapon is still (base attack + weapon dmg - 2) but the extra random2(3) penalty is removed, because it's not shown by the new monster damage description.